# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CND Rest API",
        "version": "7.0"
    },
    "servers": [
        {
            "url": "http://ins-2023.aml.fr/api/v7",
            "description": "REST APIs CND"
        },
        {
            "url": "https://ins-demo.cnd.info/api/v7",
            "description": "Site Demo"
        },
        {
            "url": "https://ins-bs.obamalou.fr/api/v7",
            "description": "Site Demo"
        }
    ],
    "paths": {
...
        "/params/0/{name}": {
            "get": {
                "tags": [
                    "Param"
                ],
                "summary": "Returns a contest parameter identified by its name",
                "operationId": "getParamByName",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "param name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "parameter",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "one parameters",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Param"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/schools": {
            "get": {
                "tags": [
                    "Schools"
                ],
                "summary": "Returns a list of schools",
                "operationId": "getAllSchools",
                "parameters": [
                    {
                        "name": "opt",
                        "in": "query",
                        "description": "options",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "usedOnly",
                                "countOnly",
                                "followUp",
                                "full"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Schools"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/schools/{id}": {
            "get": {
                "tags": [
                    "Schools"
                ],
                "operationId": "getOneSchool",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "get by Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Schools"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Response"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Param": {
                "title": "Param",
                "description": "Parameters Model",
                "properties": {
                    "param": {
                        "type": "string"
                    },
                    "valeur": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Error": {
                "title": "Error",
                "properties": {
                    "Error": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "type": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Common": {
                "title": "Common",
                "properties": {
                    "Common": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "method": {
                                    "type": "string"
                                },
                                "id": {
                                    "type": "integer"
                                },
                                "value": {
                                    "type": "string"
                                },
                                "count": {
                                    "type": "integer"
                                },
                                "type": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Header": {
                "title": "Header",
                "properties": {
                    "objectType": {
                        "type": "string"
                    },
                    "statusCode": {
                        "type": "integer"
                    },
                    "statusMessage": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Response": {
                "title": "Response",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Header"
                    },
                    {
                        "$ref": "#/components/schemas/Error"
                    },
                    {
                        "$ref": "#/components/schemas/Common"
                    }
                ]
            },
            "Schools": {
                "title": "Schools",
                "description": "School Model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "additionalAddress": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "postalCode": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "readonly": {
                        "type": "integer"
                    },
                    "validationDate": {
                        "type": "string"
                    },
                    "lastConnectionDate": {
                        "type": "string"
                    },
                    "userName": {
                        "type": "string"
                    },
                    "count": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}